:pencil: Use `autoip` as the default location
The Wunderground API is a bit weird when it comes to requesting current
conditions based on names. For some reason for most Swedish cities it
returns the result of a `geolookup` query instead when asking for
`conditions`. Using `autoip` returns the weather associated with the
(probably) location of your IP address, which tends to be correct from
my experience.