:ambulance: :pencil: Discover in OnConnect handler
The `OnConnect()` handler triggers whenever we connect to the broker,
either the first time at startup or if we're disconnected due to network
reasons. Putting the subscription to discover in the `OnConnect()`
handler ensures we always correctly subscribe since we start with a
clean session.

This makes the code a bit more messy though.